# swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"API projekt Testy Wiedzy - Aplikacje Wielowarstwowe"},"basePath":"/","paths":{"/api/Answers/{simplePollId}":{"get":{"tags":["Answers"],"operationId":"ApiAnswersBySimplePollIdGet","consumes":[],"produces":[],"parameters":[{"name":"simplePollId","in":"path","required":true,"type":"string","format":"uuid"}],"responses":{"200":{"description":"Success"}}}},"/api/Answers":{"post":{"tags":["Answers"],"operationId":"ApiAnswersPost","consumes":["application/json","text/json","application/json-patch+json"],"produces":[],"parameters":[{"name":"updateModel","in":"body","required":false,"schema":{"$ref":"#/definitions/SimplePollAnswerUpdateModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/Options":{"post":{"tags":["Options"],"operationId":"ApiOptionsPost","consumes":["application/json","text/json","application/json-patch+json"],"produces":[],"parameters":[{"name":"model","in":"body","required":false,"schema":{"$ref":"#/definitions/SimplePollOptionUpdateModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/SimplePolls/{simplePollId}":{"get":{"tags":["SimplePolls"],"operationId":"ApiSimplePollsBySimplePollIdGet","consumes":[],"produces":[],"parameters":[{"name":"simplePollId","in":"path","required":true,"type":"string","format":"uuid"}],"responses":{"200":{"description":"Success"}}}},"/api/SimplePolls":{"post":{"tags":["SimplePolls"],"operationId":"ApiSimplePollsPost","consumes":["application/json","text/json","application/json-patch+json"],"produces":[],"parameters":[{"name":"model","in":"body","required":false,"schema":{"$ref":"#/definitions/SimplePollUpdateModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/Values":{"get":{"tags":["Values"],"operationId":"ApiValuesGet","consumes":[],"produces":["text/plain","application/json","text/json"],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"type":"string"}}}}},"post":{"tags":["Values"],"operationId":"ApiValuesPost","consumes":["application/json","text/json","application/json-patch+json"],"produces":[],"parameters":[{"name":"value","in":"body","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success"}}}},"/api/Values/{id}":{"get":{"tags":["Values"],"operationId":"ApiValuesByIdGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"type":"string"}}}},"put":{"tags":["Values"],"operationId":"ApiValuesByIdPut","consumes":["application/json","text/json","application/json-patch+json"],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"value","in":"body","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["Values"],"operationId":"ApiValuesByIdDelete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}}},"definitions":{"SimplePollAnswerUpdateModel":{"type":"object","properties":{"simplePollId":{"format":"uuid","type":"string"},"simplePollOptionId":{"format":"uuid","type":"string"},"employeeName":{"type":"string"}}},"SimplePollOptionUpdateModel":{"type":"object","properties":{"simplePollId":{"format":"uuid","type":"string"},"text":{"type":"string"}}},"SimplePollUpdateModel":{"type":"object","properties":{"type":{"format":"int32","enum":[0,1],"type":"integer"},"text":{"type":"string"}}}},"securityDefinitions":{}}